Add files via upload
For this project the goal is to control a digital stepper type motor by controlling the direction of rotation and the speed control of the step in local mode by digital potentiometer and remotely with a serial usb link with the uart bridge integrated on the card fpg NEXYS 4. To do so, we will develop the local control part with digital encoders then we will add the UART communication module to load the data remotely via a pc. Finally we will use a graphical interface under python to control the stepper motor.